+++ src/components/DateRange.js
@@ -15,13 +15,13 @@ class DateRange extends Component {
     this.updatePreview = this.updatePreview.bind(this);
     this.calcNewSelection = this.calcNewSelection.bind(this);
     this.state = {
-      focusedRange: [findNextRangeIndex(props.ranges), 0],
+      focusedRange: props.initialFocusedRange || [findNextRangeIndex(props.ranges), 0],
       preview: null,
     };
     this.styles = generateStyles([coreStyles, props.classNames]);
   }
   calcNewSelection(value, isSingleValue = true) {
-    const { focusedRange } = this.state;
+    const focusedRange = this.props.focusedRange || this.state.focusedRange;
     const { ranges, onChange, maxDate, moveRangeOnFirstSelection } = this.props;
     const focusedRangeIndex = focusedRange[0];
     const selectedRange = ranges[focusedRangeIndex];
@@ -58,8 +58,8 @@ class DateRange extends Component {
     };
   }
   setSelection(value, isSingleValue) {
-    const { onChange, ranges } = this.props;
-    const { focusedRange } = this.state;
+    const { onChange, ranges, onRangeFocusChange } = this.props;
+    const focusedRange = this.props.focusedRange || this.state.focusedRange;
     const focusedRangeIndex = focusedRange[0];
     const selectedRange = ranges[focusedRangeIndex];
     if (!selectedRange) return;
@@ -74,6 +74,7 @@ class DateRange extends Component {
       focusedRange: newSelection.nextFocusRange,
       preview: null,
     });
+    onRangeFocusChange && onRangeFocusChange(newSelection.nextFocusRange);
   }
   handleRangeFocusChange(focusedRange) {
     this.setState({ focusedRange });
@@ -85,24 +86,24 @@ class DateRange extends Component {
       return;
     }
     const { rangeColors, ranges } = this.props;
-    const { focusedRange } = this.state;
+    const focusedRange = this.props.focusedRange || this.state.focusedRange;
     const color = ranges[focusedRange[0]].color || rangeColors[focusedRange[0]] || color;
     this.setState({ preview: { ...val, color } });
   }
   render() {
     return (
       <Calendar
+        focusedRange={this.state.focusedRange}
+        onRangeFocusChange={this.handleRangeFocusChange}
+        preview={this.state.preview}
+        onPreviewChange={value => {
+          this.updatePreview(value ? this.calcNewSelection(value).range : null);
+        }}
         {...this.props}
         displayMode="dateRange"
         className={classnames(this.styles.dateRangeWrapper, this.props.className)}
         onChange={this.setSelection}
-        focusedRange={this.state.focusedRange}
-        onRangeFocusChange={this.handleRangeFocusChange}
-        preview={this.state.preview}
         updateRange={val => this.setSelection(val, false)}
-        onPreviewChange={value => {
-          this.updatePreview(value ? this.calcNewSelection(value).range : null);
-        }}
         ref={target => {
           this.calendar = target;
         }}